requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib
bcrypt==4.0.1
PyMuPDF
boto3
python-multipart
